reput declaration in objects and bomb_kart and delete variable.h

+++ src/code_8006E9C0.c
@@ -173,7 +173,7 @@ void func_8006EEE8(s32 courseId) {
 // I really really don't know what D_8018D9B4 is supposed to be
 // Its value is a pointer that seems to point to compressesed data
 // of some variety (sometimes TKMK, sometimes MIO0)
-extern uintptr_t *D_8018D9B4;
+extern u8 *D_8018D9B4;
 
 void func_8006EF60(void) {
     s32 var_s2;
